Pull "ExtractConditions" out as a utility
