+++ pkg/controllers/runnable_reconciler.go
@@ -16,7 +16,6 @@ package controllers
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"reflect"
 
@@ -25,7 +24,6 @@ import (
 	rbacv1 "k8s.io/api/rbac/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/cluster-api/controllers/external"
@@ -139,7 +137,7 @@ func (r *RunnableReconciler) Reconcile(ctx context.Context, req ctrl.Request) (c
 	var trackingError error
 
 	if stampedObject != nil {
-		stampedCondition := extractConditions(stampedObject).ConditionWithType("Succeeded")
+		stampedCondition := utils.ExtractConditions(stampedObject).ConditionWithType("Succeeded")
 		if stampedCondition != nil {
 			r.conditionManager.AddPositive(conditions.StampedObjectConditionKnown(stampedCondition))
 			stampedObjectStatusPresent = true
@@ -182,23 +180,6 @@ func (r *RunnableReconciler) completeReconciliation(ctx context.Context, runnabl
 	}
 
 	return ctrl.Result{}, nil
-}
-
-// todo: dedupe from healthcheck.go
-func extractConditions(stampedObject *unstructured.Unstructured) conditions.ConditionList {
-	var conditionList conditions.ConditionList
-	maybeStatus := stampedObject.UnstructuredContent()["status"]
-	if unstructuredStatus, statusOk := maybeStatus.(map[string]interface{}); statusOk {
-		maybeConditions := unstructuredStatus["conditions"]
-		maybeConditionsJSON, err := json.Marshal(maybeConditions)
-		if err == nil {
-			err = json.Unmarshal(maybeConditionsJSON, &conditionList)
-			if err != nil {
-				return conditions.ConditionList{}
-			}
-		}
-	}
-	return conditionList
 }
 
 func (r *RunnableReconciler) trackDependencies(runnable *v1alpha1.Runnable, serviceAccountName string) {
